add a strict mode. only with strict mode labels on github that are not in config.yaml will be removed
discussed with @itrich

+++ manage.py
@@ -2,6 +2,7 @@ import os
 import github
 import logging
 import yaml
+from argparse import ArgumentParser
 
 logging.basicConfig(format='%(asctime)s - %(message)s', level=logging.INFO, datefmt='%Y-%m-%d %H:%M:%S')
 
@@ -9,6 +10,10 @@ GITHUB_TOKEN = os.environ.get("GITHUB_TOKEN")
 ORGANIZATION = os.environ.get("ORGANIZATION", "sovereigncloudstack")
 
 gh = github.Github(login_or_token=GITHUB_TOKEN)
+
+parser = ArgumentParser()
+parser.add_argument("-s", "--strict", default=False, help="Strict mode - remove labels not defined")
+args = parser.parse_args()
 
 with open("config.yaml") as fp:
     CONFIG = yaml.load(fp, Loader=yaml.SafeLoader)
@@ -51,8 +56,10 @@ for gh_repository in gh.get_organization(ORGANIZATION).get_repos(type='public'):
 
     for label in labels.keys():
         logging.info(f"{gh_repository.name} - {label} should not exist")
-        gh_label = labels[label]
-        gh_label.delete()
+        if args.strict:
+            gh_label = labels[label]
+            gh_label.delete()
+            logging.info(f"{gh_repository.name} - {label} removed")
 
     # handle milestones
